move custom queries to trait
