+++ src/Contracts/Converter/ConverterInterface.php
@@ -4,10 +4,6 @@ namespace Baka\Http\Contracts\Converter;
 
 interface ConverterInterface
 {
-    public function setCustomColumns(string $query);
-    public function setCustomTableJoins(string $query);
-    public function setCustomConditions(string $query);
-
     /**
      * Convert a Request to a whatever syntaxt we specify.
      *
