+++ src/Converter/RequestUriToSql.php
@@ -9,6 +9,7 @@ use Baka\Http\Contracts\Converter\ConverterInterface;
 use Exception;
 use Phalcon\Di;
 use Phalcon\Mvc\Model\ResultsetInterface;
+use Baka\Http\Contracts\Converter\CustomQueriesTrait;
 
 /**
  * Base QueryParser. Parse GET request for a API to a array Phalcon Model find and FindFirst can intepret.
@@ -25,6 +26,8 @@ use Phalcon\Mvc\Model\ResultsetInterface;
  */
 class RequestUriToSql extends \Phalcon\Di\Injectable implements ConverterInterface
 {
+    use CustomQueriesTrait;
+
     /**
      * @var array
      */
@@ -54,27 +57,6 @@ class RequestUriToSql extends \Phalcon\Di\Injectable implements ConverterInterfa
      * @var int
      */
     protected $offset = 25;
-
-    /**
-     * Add additional columns in search.
-     *
-     * @var string
-     */
-    protected $customColumns = null;
-
-    /**
-     * Add additional table Join
-     *
-     * @var string
-     */
-    protected $customTableJoins = null;
-
-    /**
-     * Add additional conditionals
-     *
-     * @var string
-     */
-    protected $customConditions = null;
 
     /**
      * @var array
@@ -249,40 +231,6 @@ class RequestUriToSql extends \Phalcon\Di\Injectable implements ConverterInterfa
         $rawSql['sql'] .= " LIMIT {$this->limit} OFFSET {$this->offset}";
 
         return $rawSql;
-    }
-
-    /**
-     * Set the custom columns provide by the user.
-     *
-     * @param string $query
-     * @return void
-     */
-    public function setCustomColumns(string $query) : void
-    {
-        $this->customColumns = ' ,' . $query;
-    }
-
-    /**
-     * Set the custom table by the user
-     * you can do inner joins or , table . If you are just adding a table you will need to specify the ,
-     *
-     * @param string $query
-     * @return void
-     */
-    public function setCustomTableJoins(string $query) : void
-    {
-        $this->customTableJoins = ' ' . $query;
-    }
-
-    /**
-     * set custom conditions for the query , need to start with and AND or OR
-     *
-     * @param string $query
-     * @return void
-     */
-    public function setCustomConditions(string $query) : void
-    {
-        $this->customConditions = ' ' . $query;
     }
 
     /**
